Fix #791: Refactor ClearTextBufferManager
